chore: update components' extension for constitency
